refactor classroom change naming convention from plural to singular
Change the name of the component from Classrooms to Classroom
Change the name of the controller from classroomsController to classroomController
Change the name of the model from classroomsModel to classroomModel
Change the name of the router from classroomsRouter to classroomRouter
Change the name of the file from classrooms.js to classroom.js
Update the references and imports accordingly
